Control.Monad.Freer.Fresh

Description

Composable handler for Fresh effects. This is likely to be of use when implementing De Bruijn naming/scopes.

Using http://okmij.org/ftp/Haskell/extensible/Eff1.hs as a starting point.

Documentation

data Fresh r where

Fresh effect model.

Constructors

Fresh :: Fresh Int

fresh :: Member Fresh effs => Eff effs Int

Request a fresh effect.

runFresh :: Int -> Eff (Fresh ': effs) a -> Eff effs (a, Int)

Handler for Fresh effects, with an Int for a starting value. The return value includes the next fresh value.

evalFresh :: Int -> Eff (Fresh ': effs) a -> Eff effs a

Handler for Fresh effects, with an Int for a starting value. Discards the next fresh value.